modularize all routing, some snippets
