+++ lib/tagMap.js
@@ -1,0 +1,15 @@
+const tagMap = new Map();
+
+tagMap.set('.infogather', 'Information Gathering');
+tagMap.set('.analysis', 'Problem Analysis');
+tagMap.set('.opportunity', 'Opportunity Analysis');
+tagMap.set('.ideation', 'Ideation / Evaluation');
+tagMap.set('.stratdev', 'Strategy Development');
+tagMap.set('.busstrat', 'Business Strategy');
+tagMap.set('.expstrat', 'Experience Strategy');
+tagMap.set('.stratfore', 'Strategic Foresight');
+tagMap.set('.coretool', 'Core Tool');
+tagMap.set('.secondary', 'Secondary Tool');
+tagMap.set('.collaboration', 'Client collaboration');
+
+module.exports = tagMap;
