+++ views/header.ejs
@@ -12,15 +12,15 @@
   <!-- Bootstrap CSS -->
   <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
   <!-- local CSS -->
-  <link rel="stylesheet" href="css/styles.css">
+  <link rel="stylesheet" href="/css/styles.css">
   <title>Strategy Toolkit</title>
 </head>
 
 <body>
   <header>
     <nav class="navbar navbar-expand-md navbar-dark">
-      <a class="navbar-brand navbar-logo" href="/strategytoolkit">
-        <img src="images/toolkit_logo.png" alt="toolkit logo">
+      <a class="navbar-brand navbar-logo" href="/">
+        <img src="/images/toolkit_logo.png" alt="toolkit logo">
       </a>
       <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
         <span class="navbar-toggler-icon"></span>
@@ -28,13 +28,13 @@
       <div class="collapse navbar-collapse" id="navbarSupportedContent">
         <ul class="navbar-nav ml-auto">
           <li class="nav-item">
-            <a class="nav-link <%= page_name === 'about' && 'active' %>" href="/strategytoolkit/about">About</a>
+            <a class="nav-link <%= page_name === 'about' && 'active' %>" href="/about">About</a>
           </li>
           <li class="nav-item">
-            <a class="nav-link <%= page_name === 'addtool' && 'active' %>" href="/strategytoolkit/addtool">Add Tool</a>
+            <a class="nav-link <%= page_name === 'addtool' && 'active' %>" href="/addtool">Add Tool</a>
           </li>
           <li class="nav-item">
-            <a class="nav-link display="none" <%= page_name === 'tool' && 'active' %>" href="/strategytoolkit/addtool"></a>
+            <a class="nav-link display="none" <%= page_name === 'tool' && 'active' %>" href="/addtool"></a>
           </li>
         </ul>
       </div>
